switched to .net8, bumped dbscan version to 3.0.0, added TrackablePointT, CalculateClusterT. Aligned main.yml and nuspec.
