+++ .github/workflows/main.yml
@@ -5,23 +5,24 @@ on:
       - master
     paths-ignore:
       - README.md
+  workflow_dispatch:
 
 jobs:
   build:
     runs-on: windows-latest
     steps:
-    - name: Git Checkout
-      uses: actions/checkout@master
-    
-    - name: Setup MSBuild.exe
-      uses: microsoft/setup-msbuild@v1.0.0
+      - name: Git Checkout
+        uses: actions/checkout@master
 
-    - name: Setup Nuget.exe
-      uses: nuget/setup-nuget@v1
-      
-    - name: Publish VL Nuget
-      uses: vvvv/PublishVLNuget@1.0.28
-      with:
-        csproj: src/DBSCAN.Utils/DBSCAN.Utils.csproj
-        nuspec: deployment/VL.DBSCAN.nuspec
-        nuget-key: ${{ secrets.NUGET_KEY }}
+      - name: Setup MSBuild.exe
+        uses: microsoft/setup-msbuild@v1.0.0
+
+      - name: Setup Nuget.exe
+        uses: nuget/setup-nuget@v1
+
+      - name: Publish VL Nuget
+        uses: vvvv/PublishVLNuget@1.0.28
+        with:
+          csproj: src/VL.Dbscan.csproj
+          nuspec: deployment/VL.DBSCAN.nuspec
+          nuget-key: ${{ secrets.NUGET_KEY }}
