--- src/DBSCAN.Utils/DBSCAN.Utils.csproj
@@ -1,23 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <LangVersion>7.2</LangVersion>
-    <Configurations>Debug;Release;VL</Configurations>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <OutputPath></OutputPath>
-    <DocumentationFile>D:\Documents\vvvv\VL\VL.DBSCAN\vs\DBSCAN.Utils\DBSCAN.Utils.xml</DocumentationFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <OutputPath></OutputPath>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="DBSCAN" Version="2.0.12" />
-    <PackageReference Include="Stride.Core.Mathematics" Version="4.0.0.1-beta04-1265" />
-  </ItemGroup>
-
-</Project>
