--- src/DBSCAN.Utils/DBSCAN.Utils.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0"?>
-<doc>
-    <assembly>
-        <name>DBSCAN.Utils</name>
-    </assembly>
-    <members>
-        <member name="T:DBSCAN.Utils.TrackablePoint">
-            <summary>
-            Class implementing IPointData to use VL's Vector2 with DBSCAN plugin
-            </summary>
-        </member>
-        <member name="M:DBSCAN.Utils.TrackablePoint.#ctor(SharpDX.Vector2)">
-            <summary>
-            Creates a TrackablePoint from a Vector2 to be used with VL DBSCAN plugin
-            </summary>
-            <param name="vector"></param>
-        </member>
-        <member name="P:DBSCAN.Utils.TrackablePoint.Point">
-            <summary>
-            DBSCAN library's point
-            </summary>
-        </member>
-        <member name="M:DBSCAN.Utils.TrackablePoint.ToVector2(DBSCAN.Utils.TrackablePoint)">
-            <summary>
-            Converts a TrackablePoint to a Vector2 for further use in VL
-            </summary>
-            <param name="point"></param>
-            <returns></returns>
-        </member>
-    </members>
-</doc>
