--- src/DBSCAN.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2046
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBSCAN.Utils", "DBSCAN.Utils\DBSCAN.Utils.csproj", "{8225A298-ADB9-4EB2-A74D-8F49C63390B7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-		VL|Any CPU = VL|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.VL|Any CPU.ActiveCfg = Release|Any CPU
-		{8225A298-ADB9-4EB2-A74D-8F49C63390B7}.VL|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {36E5013F-EF1A-458C-9434-8C77C445D1D1}
-	EndGlobalSection
-EndGlobal
