+++ src/VL.Dbscan.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35507.96
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VL.Dbscan", "VL.Dbscan.csproj", "{321F6E40-B93F-EB5F-B75B-648D14A51ED6}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{321F6E40-B93F-EB5F-B75B-648D14A51ED6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{321F6E40-B93F-EB5F-B75B-648D14A51ED6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{321F6E40-B93F-EB5F-B75B-648D14A51ED6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{321F6E40-B93F-EB5F-B75B-648D14A51ED6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {3C2AE5CE-B972-4282-BD6D-D2EFB573E2E1}
+	EndGlobalSection
+EndGlobal
